refactor: Remove ModelBuilder abstraction and simplify entity configuration.
