+++ src/TechNovaLab.Irrigo.Infrastructure/TechNovaLab.Irrigo.Infrastructure.csproj
@@ -8,14 +8,11 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.EntityFrameworkCore" Version="9.0.0" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="9.0.0" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\TechNovaLab.Irrigo.Application\TechNovaLab.Irrigo.Application.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Database\Configurations\" />
   </ItemGroup>
 
 </Project>
