--- src/TechNovaLab.Irrigo.SharedKernel/Abstractions/Data/IModelBuilder.cs
@@ -1,9 +1,0 @@
-﻿using TechNovaLab.Irrigo.SharedKernel.Core;
-
-namespace TechNovaLab.Irrigo.SharedKernel.Abstractions.Data
-{
-    public interface IModelBuilder
-    {
-        IEntityTypeConfigurationBuilder<TEntity> Entity<TEntity>() where TEntity : EntityBase;
-    }
-}
